refactor: move WfcIntent.java to app package
